Conflicts:
	IMBApertureParser.m
	IMBAppleMediaParser.h
	IMBAppleMediaParser.m
	IMBAudioObjectViewController.m
	IMBComboTextCell.m
	IMBFlickrNode.h
	IMBFlickrNode.m
	IMBFlickrParser.m
	IMBImageObjectViewController.m
	IMBLinkObjectViewController.m
	IMBMetadataTransformer.m
	IMBNode.m
	IMBObjectViewController.h
	IMBObjectViewController.m
	IMBOutlineView.m
	IMBTestAppDelegate.m
	IMBTimecodeTransformer.m
	IMBiPhotoParser.m
	NSData+SKExtensions.m
	NSFileManager+iMedia.m
	iMedia.xcodeproj/project.pbxproj

+++ NSFileManager+iMedia.m
@@ -301,7 +301,7 @@
     NSString *result = [path stringByAppendingPathComponent:filename];
     NSInteger i = 1;
     while ([self fileExistsAtPath:result]) {
-        filename = [NSString stringWithFormat:@"%@ %ld.%@", basename, (long)i, extension];
+        filename = [NSString stringWithFormat:@"%@ %ld.%@", basename, i, extension];
         result = [path stringByAppendingPathComponent:filename];
         i++;
     }
